refactor: JpaRepository 패키지 이동
