+++ src/main/java/kitchenpos/order/domain/repository/JpaOrderRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.order.infra;
+package kitchenpos.order.domain.repository;
 
 import java.util.UUID;
 import kitchenpos.order.domain.model.Order;
-import kitchenpos.order.domain.repository.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
