+++ src/main/java/kitchenpos/order/domain/repository/JpaOrderTableRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.order.infra;
+package kitchenpos.order.domain.repository;
 
 import java.util.UUID;
 import kitchenpos.order.domain.model.OrderTable;
-import kitchenpos.order.domain.repository.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
