+++ src/main/java/kitchenpos/product/domain/repository/JpaProductRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.product.infra;
+package kitchenpos.product.domain.repository;
 
 import java.util.UUID;
 import kitchenpos.product.domain.model.Product;
-import kitchenpos.product.domain.repository.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
